Changing all the documentation
