--- docs/README.md
@@ -1,26 +1,0 @@
-<div align="center">
-  <a href="https://www.pokt.network">
-    <img src="https://user-images.githubusercontent.com/2219004/151564884-212c0e40-3bfa-412e-a341-edb54b5f1498.jpeg" alt="Pocket Network logo" width="340"/>
-  </a>
-</div>
-
-# Pocket V1 Documentation
-
-## Sections
-
-- [Contributing Guide](contributing/CONTRIBUTING.md)
-- [Dependencies Information](deps/README.md)
-- [Testing Information](testing/README.md)
-
-## Support & Contact
-
-<div>
-  <a  href="https://twitter.com/poktnetwork" ><img src="https://img.shields.io/twitter/url/http/shields.io.svg?style=social"></a>
-  <a href="https://t.me/POKTnetwork"><img src="https://img.shields.io/badge/Telegram-blue.svg"></a>
-  <a href="https://www.facebook.com/POKTnetwork" ><img src="https://img.shields.io/badge/Facebook-red.svg"></a>
-  <a href="https://research.pokt.network"><img src="https://img.shields.io/discourse/https/research.pokt.network/posts.svg"></a>
-</div>
-
-## License
-
-This project is licensed under the MIT License; see the [LICENSE.md](LICENSE.md) file for details.
