--- docs/SUMMARY.md
@@ -1,4 +1,0 @@
-# Table of contents
-
-* [Welcome](README.md)
-* [How to Contribute](CONTRIBUTING.md)
